Sync db changes with offline branch and extract common entities.
