+++ libs/pandautils/src/main/java/com/instructure/pandautils/di/DatabaseModule.kt
@@ -2,6 +2,10 @@ package com.instructure.pandautils.di
 
 import com.instructure.pandautils.room.appdatabase.AppDatabase
 import com.instructure.pandautils.room.appdatabase.daos.*
+import com.instructure.pandautils.room.common.daos.AttachmentDao
+import com.instructure.pandautils.room.common.daos.AuthorDao
+import com.instructure.pandautils.room.common.daos.MediaCommentDao
+import com.instructure.pandautils.room.common.daos.SubmissionCommentDao
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
